+++ whitelabel/models/src/main/scala/it/agilelab/bigdata/wasp/whitelabel/models/example/ExamplePipegraphModel.scala
@@ -4,7 +4,7 @@ import it.agilelab.bigdata.wasp.core.models._
 
 private[wasp] object ExamplePipegraphModel {
 
-  def apply() = PipegraphModel(
+  lazy val pipegraph = PipegraphModel(
     name = "ExamplePipegraph",
     description = "Description of Example Pipegraph",
     owner = "user",
@@ -17,18 +17,18 @@ private[wasp] object ExamplePipegraphModel {
         name = "Write on console",
         inputs = List(
           ReaderModel.kafkaReader(
-            ExampleTopicModel.topic_name,
-            ExampleTopicModel.topic_name
+            ExampleTopicModel.topic.name,
+            ExampleTopicModel.topic.name
           )
         ),
         output = WriterModel.consoleWriter("console-writer"),
         mlModels = List.empty,
         strategy = None,
         kafkaAccessType = LegacyStreamingETLModel.KAFKA_ACCESS_TYPE_RECEIVED_BASED,
-        config = Map.empty
+        config = Map()
       )
     ),
-    rtComponents = Nil,
+    rtComponents = List(),
 
     dashboard = None,
     isActive = false)