+++ whitelabel/models/src/main/scala/it/agilelab/bigdata/wasp/whitelabel/models/test/TestProducerModel.scala
@@ -4,27 +4,24 @@ import it.agilelab.bigdata.wasp.core.models.ProducerModel
 
 private[wasp] object TestProducerModel {
 
-  object JSON {
-    def apply() = ProducerModel(
+  lazy val json = ProducerModel(
       name = "TestJSONProducer",
       className = "it.agilelab.bigdata.wasp.whitelabel.producers.TestProducerGuardian",
-      topicName = Some(TestTopicModel.testJsonTopic.name),
+      topicName = Some(TestTopicModel.json.name),
       isActive = false,
       configuration = None,
       isRemote = false,
       isSystem = false
     )
-  }
 
-  object AVRO {
-    def apply() = ProducerModel(
+  lazy val avro  = ProducerModel(
       name = "TestAVROProducer",
       className = "it.agilelab.bigdata.wasp.whitelabel.producers.TestProducerGuardian",
-      topicName = Some(TestTopicModel.testAvroTopic.name),
+      topicName = Some(TestTopicModel.avro.name),
       isActive = false,
       configuration = None,
       isRemote = false,
       isSystem = false
     )
-  }
+
 }