+++ whitelabel/models/src/main/scala/it/agilelab/bigdata/wasp/whitelabel/models/test/TestRawModel.scala
@@ -7,30 +7,28 @@ import org.apache.spark.sql.types._
 private[wasp] object TestRawModel {
 
   /* for Pipegraph */
-  lazy val nestedSchemaRawModel = RawModel(name = "TestRawNestedSchemaModel",
+  lazy val nested = RawModel(name = "TestRawNestedSchemaModel",
     uri="hdfs://namenode:9000/user/test_nested/",
     timed = true,
-    schema = nestedSchema.json)
-  private lazy val nestedSchema = StructType(Seq(
-    StructField("id", StringType),
-    StructField("number", IntegerType),
-    StructField("nested", StructType(Seq(
-      StructField("field1", StringType),
-      StructField("field2", LongType),
-      StructField("field3", StringType)
-    )))
-  ))
+    schema = StructType(Seq(
+      StructField("id", StringType),
+      StructField("number", IntegerType),
+      StructField("nested", StructType(Seq(
+        StructField("field1", StringType),
+        StructField("field2", LongType),
+        StructField("field3", StringType)
+      )))
+    )).json)
 
   /* for BatchJob */
-  lazy val flatSchemaRawModel = RawModel(name = "TestRawFlatSchemaModel",
+  lazy val flat = RawModel(name = "TestRawFlatSchemaModel",
     uri="hdfs://namenode:9000/user/test_flat/",
     timed = true,
-    schema = flatSchema.json)
-  private lazy val flatSchema = StructType(Seq(
-    StructField("id", StringType),
-    StructField("number", LongType),
-    StructField("nested.field1", StringType),
-    StructField("nested.field2", LongType),
-    StructField("nested.field3", StringType)
-  ))
-}
+    schema = StructType(Seq(
+      StructField("id", StringType),
+      StructField("number", LongType),
+      StructField("nested.field1", StringType),
+      StructField("nested.field2", LongType),
+      StructField("nested.field3", StringType)
+    )).json)
+  }